+++ include/ignition/gazebo/components/Magnetometer.hh
@@ -17,13 +17,15 @@
 #ifndef IGNITION_GAZEBO_COMPONENTS_MAGNETOMETER_HH_
 #define IGNITION_GAZEBO_COMPONENTS_MAGNETOMETER_HH_
 
-#include <sdf/Element.hh>
+#include <ignition/msgs/sensor.pb.h>
+#include <sdf/Sensor.hh>
 
 #include <ignition/gazebo/config.hh>
 #include <ignition/gazebo/Export.hh>
 
 #include <ignition/gazebo/components/Factory.hh>
-#include "ignition/gazebo/components/Component.hh"
+#include <ignition/gazebo/components/Component.hh>
+#include <ignition/gazebo/Conversions.hh>
 
 namespace ignition
 {
@@ -33,9 +35,41 @@ namespace gazebo
 inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-  /// \brief TODO(anyone) Substitute with sdf::Magnetometer once that exists?
-  /// This is currently the whole <sensor> element.
-  using Magnetometer = Component<sdf::ElementPtr, class MagnetometerTag>;
+  /// \brief A component type that contains a magnetometer sensor,
+  /// sdf::Magnetometer, information.
+  using MagnetometerBase = Component<sdf::Sensor, class MagnetometerTag>;
+
+  /// \brief This component holds a magnetometer sensor.
+  class Magnetometer : public MagnetometerBase
+  {
+    // Documentation inherited
+    public: Magnetometer() : MagnetometerBase()
+    {
+    }
+
+    // Documentation inherited
+    public: explicit Magnetometer(const sdf::Sensor &_data)
+      : MagnetometerBase(_data)
+    {
+    }
+
+    // Documentation inherited
+    public: void Serialize(std::ostream &_out) const override
+    {
+      auto msg = convert<msgs::Sensor>(this->Data());
+      msg.SerializeToOstream(&_out);
+    }
+
+    // Documentation inherited
+    public: void Deserialize(std::istream &_in) override
+    {
+      msgs::Sensor msg;
+      msg.ParseFromIstream(&_in);
+
+      this->Data() = convert<sdf::Sensor>(msg);
+    }
+  };
+
   IGN_GAZEBO_REGISTER_COMPONENT(
       "ign_gazebo_components.Magnetometer", Magnetometer)
 }